// client/src/screens/Robot/Enums.ts
export enum MemberState {
  AtRest = 0,
  SlightlyContracted = 1,
  Contracted = 2,
  StronglyContracted = 3,
}

export enum MemberRotation {
  Negative90 = 0,
  Negative45 = 1,
  AtRest = 2,
  Positive45 = 3,
  Positive90 = 4,
  Positive135 = 5,
  Positive180 = 6,
}

export enum MemberInclination {
  Upward = 0,
  AtRest = 1,
  Downward = 2,
}